Move level role commands into LevelRolesCommand
This commit moves the commands to manage level role rewards into a single slash command
