+++ README.md
@@ -26,16 +26,16 @@ The bot has many slash commands - the amazing discord feature. Here is the list 
 
 |Command|Description|Permission|
 |------------|---|---|
-|`/add-level-role <role> <level>`| Adds a role as a reward for a level | `MANAGE_ROLES` |
-|`/remove-level-role <role>` | Removes a role from any reward level | `MANAGE_ROLES` |
-|`/get-level-roles`| Gets a list of reward roles |  |
+|`/level-roles add <role> <level>`| Adds a role as a reward for a level | `MANAGE_ROLES`, `MANAGE_SERVER` |
+|`/level-roles remove <role>` | Removes a role from any reward level | `MANAGE_ROLES`, `MANAGE_SERVER` |
+|`/level-roles list`| Gets a list of reward roles | `MANAGE_ROLES`, `MANAGE_SERVER` |
 |`/levels get [member]`| Get your own or another member's level | |
 |`/levels top`| View the top 5 levels | |
 |`/levels set <member> <level>`| Set a member's level | `ADMINISTRATOR` |
 |`/warnings add <member> [reason]`| Give a member a warning | `MODERATE_MEMBERS` |
-|`/warnings remove <member> <id>`| Remove a warning from a member | `MODERATE_MEMBERS` |
+|`/warnings remove <member> <id>`| Remove a warning from a member | `ADMINISTRATOR` |
 |`/warnings list <member>`| View the warnings of a member | `MODERATE_MEMBERS` |
-|`/warnings clear <member>`| Clear a member's warnings | `MODERATE_MEMBERS` |
+|`/warnings clear <member>`| Clear a member's warnings | `ADMINISTRATOR` |
 |`/logging set-channel <channel>`| Set the bot logging channel | `MANAGE_SERVER` |
 |`/logging remove-channel <channel>`| Remove the bot logging channel | `MANAGE_SERVER` |
 |`/welcome-message set-channel <channel>`| Set the welcome message channel | `MANAGE_SERVER` |
