+++ src/main/java/me/grayingout/bot/commands/BotCommandManager.java
@@ -1,12 +1,10 @@
 package me.grayingout.bot.commands;
 
-import me.grayingout.bot.commands.implementations.AddLevelRoleCommand;
 import me.grayingout.bot.commands.implementations.BulkDeleteCommand;
-import me.grayingout.bot.commands.implementations.GetLevelRolesCommand;
 import me.grayingout.bot.commands.implementations.HelloCommand;
+import me.grayingout.bot.commands.implementations.LevelRolesCommand;
 import me.grayingout.bot.commands.implementations.LevelsCommand;
 import me.grayingout.bot.commands.implementations.LoggingCommand;
-import me.grayingout.bot.commands.implementations.RemoveLevelRoleCommand;
 import me.grayingout.bot.commands.implementations.RulesCommand;
 import me.grayingout.bot.commands.implementations.SlowmodeCommand;
 import me.grayingout.bot.commands.implementations.WarningsCommand;
@@ -54,17 +52,7 @@ public final class BotCommandManager {
     /**
      * The {@code /add-level-role <role:role> <level:int>} command
      */
-    public static final AddLevelRoleCommand ADD_LEVEL_ROLE_COMMAND = new AddLevelRoleCommand();
-
-    /**
-     * The {@code /get-level-roles} command
-     */
-    public static final GetLevelRolesCommand GET_LEVEL_ROLES_COMMAND = new GetLevelRolesCommand();
-
-    /**
-     * The {@code /remove-level-role} command
-     */
-    public static final RemoveLevelRoleCommand REMOVE_LEVEL_ROLE_COMMAND = new RemoveLevelRoleCommand();
+    public static final LevelRolesCommand LEVEL_ROLES_COMMAND = new LevelRolesCommand();
 
     /**
      * The {@code /welcome-message (set-channel <channel:channel> | remove-channel | set-message <message:string>)} command
